+++ sipXmediaLib/include/rtcp/RTCPTimer.h
@@ -276,11 +276,11 @@ private:    // Private Methods
  *
  *
  */
-    static void ReportingAlarm(timer_t tTimer, int iArgument);
+    static void ReportingAlarm(timer_t tTimer, intptr_t iArgument);
 
 #elif defined(__pingtel_on_posix__) /* ] [ */
 
-    static void ReportingAlarm(const int userData, const int eventData);
+    static void ReportingAlarm(const intptr_t userData, const intptr_t eventData);
 
 #endif /* ] */
